# app/api/openapi/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/geocoding/addresses": {
      "post": {
        "description": "Port of `@InvocableMethod GeocodingService.geocodeAddresses` (called from the Create_property flow). Returns one Coordinates entry per input address, in order.",
        "operationId": "geocodeAddresses",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GeocodeAddressesDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CoordinatesDto"
                  }
                }
              }
            }
          },
          "501": {
            "description": "Not ported yet (UNT3-17)"
          }
        },
        "summary": "Geocode one or more addresses",
        "tags": [
          "geocoding"
        ]
      }
    },
...
  },
  "info": {
    "title": "Dreamhouse API",
    "description": "Dreamhouse API — NestJS + Prisma on PostgreSQL; target of the Salesforce (dreamhouse-lwc) migration",
    "version": "0.1.0",
    "contact": {},
    "license": {
      "name": "CC0-1.0",
      "url": "https://creativecommons.org/publicdomain/zero/1.0/"
    }
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "GeocodingAddressDto": {
        "type": "object",
        "properties": {
          "street": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "state": {
            "type": "string"
          },
          "country": {
            "type": "string"
          },
          "postalcode": {
            "type": "string"
          }
        }
      },
      "GeocodeAddressesDto": {
        "type": "object",
        "properties": {
          "addresses": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GeocodingAddressDto"
            }
          }
        },
        "required": [
          "addresses"
        ]
      },
      "CoordinatesDto": {
        "type": "object",
        "properties": {
          "lat": {
            "type": "object",
            "nullable": true
          },
          "lon": {
            "type": "object",
            "nullable": true
          }
        }
      },
...
    }
  }
}